+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - /method/{uid}/verify no longer requires authentication
 - Validation attempt on expired reset now issues a new reset
 - Password change now clears out the auth token if `auth_type` is reset
+- Only provide spouse and manager recovery methods if the user
+  has not added and verified others.
 ### Removed
 - Removed support for phone password recovery methods
 - Removed option to use local emailer. External email service is now required.